Split the project parsing and the project into separate classes - Added ProjectReader
