+++ src/Microsoft.Dnx.Runtime/Project.cs
@@ -3,13 +3,9 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Runtime.Versioning;
-using System.Text;
-using Microsoft.Dnx.Runtime.Helpers;
-using Microsoft.Dnx.Runtime.Json;
 using NuGet;
 
 namespace Microsoft.Dnx.Runtime
@@ -23,19 +19,19 @@ namespace Microsoft.Dnx.Runtime
 
         public static TypeInformation DefaultCompiler = DefaultRuntimeCompiler;
 
-        private readonly Dictionary<FrameworkName, TargetFrameworkInformation> _targetFrameworks = new Dictionary<FrameworkName, TargetFrameworkInformation>();
-        private readonly Dictionary<FrameworkName, CompilerOptions> _compilationOptions = new Dictionary<FrameworkName, CompilerOptions>();
-        private readonly Dictionary<string, CompilerOptions> _configurations = new Dictionary<string, CompilerOptions>(StringComparer.OrdinalIgnoreCase);
+        // REVIEW: It's kinda hacky making these internal but the reader needs to set them
+        internal Dictionary<FrameworkName, TargetFrameworkInformation> _targetFrameworks = new Dictionary<FrameworkName, TargetFrameworkInformation>();
+        internal Dictionary<FrameworkName, CompilerOptions> _compilerOptionsByFramework = new Dictionary<FrameworkName, CompilerOptions>();
+        internal Dictionary<string, CompilerOptions> _compilerOptionsByConfiguration = new Dictionary<string, CompilerOptions>(StringComparer.OrdinalIgnoreCase);
 
-        private CompilerOptions _defaultCompilerOptions;
-
-        private TargetFrameworkInformation _defaultTargetFrameworkConfiguration;
+        internal CompilerOptions _defaultCompilerOptions;
+        internal TargetFrameworkInformation _defaultTargetFrameworkConfiguration;
 
         public Project()
         {
         }
 
-        public string ProjectFilePath { get; private set; }
+        public string ProjectFilePath { get; set; }
 
         public string ProjectDirectory
         {
@@ -45,11 +41,11 @@ namespace Microsoft.Dnx.Runtime
             }
         }
 
-        public string Name { get; private set; }
+        public string Name { get; set; }
 
         public string Title { get; set; }
 
-        public string Description { get; private set; }
+        public string Description { get; set; }
 
         public string Copyright { get; set; }
 
@@ -59,39 +55,39 @@ namespace Microsoft.Dnx.Runtime
 
         public string ReleaseNotes { get; set; }
 
-        public string[] Authors { get; private set; }
+        public string[] Authors { get; set; }
 
-        public string[] Owners { get; private set; }
+        public string[] Owners { get; set; }
 
-        public IDictionary<string, string> Repository { get; private set; }
+        public IDictionary<string, string> Repository { get; set; }
 
         public bool EmbedInteropTypes { get; set; }
 
-        public SemanticVersion Version { get; private set; }
+        public SemanticVersion Version { get; set; }
 
-        public Version AssemblyFileVersion { get; private set; }
+        public Version AssemblyFileVersion { get; set; }
 
-        public IList<LibraryDependency> Dependencies { get; private set; }
+        public IList<LibraryDependency> Dependencies { get; set; }
 
-        public CompilerServices CompilerServices { get; private set; }
+        public CompilerServices CompilerServices { get; set; }
 
-        public string WebRoot { get; private set; }
+        public string WebRoot { get; set; }
 
-        public string EntryPoint { get; private set; }
+        public string EntryPoint { get; set; }
 
-        public string ProjectUrl { get; private set; }
+        public string ProjectUrl { get; set; }
 
         public string LicenseUrl { get; set; }
 
         public string IconUrl { get; set; }
 
-        public bool RequireLicenseAcceptance { get; private set; }
+        public bool RequireLicenseAcceptance { get; set; }
 
-        public string[] Tags { get; private set; }
+        public string[] Tags { get; set; }
 
         public bool IsLoadable { get; set; }
 
-        public ProjectFilesCollection Files { get; private set; }
+        public ProjectFilesCollection Files { get; set; }
 
         public IDictionary<string, string> Commands { get; } = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
 
@@ -104,7 +100,7 @@ namespace Microsoft.Dnx.Runtime
 
         public IEnumerable<string> GetConfigurations()
         {
-            return _configurations.Keys;
+            return _compilerOptionsByConfiguration.Keys;
         }
 
         public static bool HasProjectFile(string path)
@@ -142,7 +138,8 @@ namespace Microsoft.Dnx.Runtime
             {
                 using (var stream = File.OpenRead(projectPath))
                 {
-                    project = GetProjectFromStream(stream, projectName, projectPath, diagnostics);
+                    var reader = new ProjectReader();
+                    project = reader.ReadProject(stream, projectName, projectPath, diagnostics);
                 }
             }
             catch (Exception ex)
@@ -153,314 +150,8 @@ namespace Microsoft.Dnx.Runtime
             return true;
         }
 
-        internal static Project GetProjectFromStream(Stream stream, string projectName, string projectPath, ICollection<DiagnosticMessage> diagnostics = null)
-        {
-            var project = new Project();
-
-            var reader = new StreamReader(stream);
-            var rawProject = JsonDeserializer.Deserialize(reader) as JsonObject;
-            if (rawProject == null)
-            {
-                throw FileFormatException.Create(
-                    "The JSON file can't be deserialized to a JSON object.",
-                    projectPath);
-            }
-
-            // Meta-data properties
-            project.Name = projectName;
-            project.ProjectFilePath = Path.GetFullPath(projectPath);
-
-            var version = rawProject.Value("version") as JsonString;
-            if (version == null)
-            {
-                project.Version = new SemanticVersion("1.0.0");
-            }
-            else
-            {
-                try
-                {
-                    var buildVersion = Environment.GetEnvironmentVariable("DNX_BUILD_VERSION");
-                    project.Version = SpecifySnapshot(version, buildVersion);
-                }
-                catch (Exception ex)
-                {
-                    throw FileFormatException.Create(ex, version, project.ProjectFilePath);
-                }
-            }
-
-            var fileVersion = Environment.GetEnvironmentVariable("DNX_ASSEMBLY_FILE_VERSION");
-            if (string.IsNullOrWhiteSpace(fileVersion))
-            {
-                project.AssemblyFileVersion = project.Version.Version;
-            }
-            else
-            {
-                try
-                {
-                    var simpleVersion = project.Version.Version;
-                    project.AssemblyFileVersion = new Version(simpleVersion.Major,
-                        simpleVersion.Minor,
-                        simpleVersion.Build,
-                        int.Parse(fileVersion));
-                }
-                catch (FormatException ex)
-                {
-                    throw new FormatException("The assembly file version is invalid: " + fileVersion, ex);
-                }
-            }
-
-            project.Description = rawProject.ValueAsString("description");
-            project.Summary = rawProject.ValueAsString("summary");
-            project.Copyright = rawProject.ValueAsString("copyright");
-            project.Title = rawProject.ValueAsString("title");
-            project.WebRoot = rawProject.ValueAsString("webroot");
-            project.EntryPoint = rawProject.ValueAsString("entryPoint");
-            project.ProjectUrl = rawProject.ValueAsString("projectUrl");
-            project.LicenseUrl = rawProject.ValueAsString("licenseUrl");
-            project.IconUrl = rawProject.ValueAsString("iconUrl");
-
-            project.Authors = rawProject.ValueAsStringArray("authors") ?? new string[] { };
-            project.Owners = rawProject.ValueAsStringArray("owners") ?? new string[] { };
-            project.Tags = rawProject.ValueAsStringArray("tags") ?? new string[] { };
-
-            project.Language = rawProject.ValueAsString("language");
-            project.ReleaseNotes = rawProject.ValueAsString("releaseNotes");
-
-            project.RequireLicenseAcceptance = rawProject.ValueAsBoolean("requireLicenseAcceptance", defaultValue: false);
-            project.IsLoadable = rawProject.ValueAsBoolean("loadable", defaultValue: true);
-            // TODO: Move this to the dependencies node
-            project.EmbedInteropTypes = rawProject.ValueAsBoolean("embedInteropTypes", defaultValue: false);
-
-            project.Dependencies = new List<LibraryDependency>();
-
-            // Project files
-            project.Files = new ProjectFilesCollection(rawProject, project.ProjectDirectory, project.ProjectFilePath);
-
-            var compilerInfo = rawProject.ValueAsJsonObject("compiler");
-            if (compilerInfo != null)
-            {
-                var languageName = compilerInfo.ValueAsString("name") ?? "C#";
-                var compilerAssembly = compilerInfo.ValueAsString("compilerAssembly");
-                var compilerType = compilerInfo.ValueAsString("compilerType");
-
-                var compiler = new TypeInformation(compilerAssembly, compilerType);
-                project.CompilerServices = new CompilerServices(languageName, compiler);
-            }
-
-            var commands = rawProject.Value("commands") as JsonObject;
-            if (commands != null)
-            {
-                foreach (var key in commands.Keys)
-                {
-                    var value = commands.ValueAsString(key);
-                    if (value != null)
-                    {
-                        project.Commands[key] = value;
-                    }
-                }
-            }
-
-            var scripts = rawProject.Value("scripts") as JsonObject;
-            if (scripts != null)
-            {
-                foreach (var key in scripts.Keys)
-                {
-                    var stringValue = scripts.ValueAsString(key);
-                    if (stringValue != null)
-                    {
-                        project.Scripts[key] = new string[] { stringValue };
-                        continue;
-                    }
-
-                    var arrayValue = scripts.ValueAsStringArray(key);
-                    if (arrayValue != null)
-                    {
-                        project.Scripts[key] = arrayValue;
-                        continue;
-                    }
-
-                    throw FileFormatException.Create(
-                        string.Format("The value of a script in {0} can only be a string or an array of strings", ProjectFileName),
-                        scripts.Value(key),
-                        project.ProjectFilePath);
-                }
-            }
-
-            var repository = rawProject.Value("repository") as JsonObject;
-            if (repository != null)
-            {
-                project.Repository = repository
-                    .Keys
-                    .ToDictionary(
-                        key => key,
-                        key => repository.ValueAsString(key).Value);
-            }
-
-            project.BuildTargetFrameworksAndConfigurations(rawProject, diagnostics);
-
-            PopulateDependencies(
-                project.ProjectFilePath,
-                project.Dependencies,
-                rawProject,
-                "dependencies",
-                isGacOrFrameworkReference: false);
-
-            return project;
-        }
-
-        private static SemanticVersion SpecifySnapshot(string version, string snapshotValue)
-        {
-            if (version.EndsWith("-*"))
-            {
-                if (string.IsNullOrEmpty(snapshotValue))
-                {
-                    version = version.Substring(0, version.Length - 2);
-                }
-                else
-                {
-                    version = version.Substring(0, version.Length - 1) + snapshotValue;
-                }
-            }
-
-            return new SemanticVersion(version);
-        }
-
-        private static void PopulateDependencies(
-            string projectPath,
-            IList<LibraryDependency> results,
-            JsonObject settings,
-            string propertyName,
-            bool isGacOrFrameworkReference)
-        {
-            var dependencies = settings.ValueAsJsonObject(propertyName);
-            if (dependencies != null)
-            {
-                foreach (var dependencyKey in dependencies.Keys)
-                {
-                    if (string.IsNullOrEmpty(dependencyKey))
-                    {
-                        throw FileFormatException.Create(
-                            "Unable to resolve dependency ''.",
-                            dependencies.Value(dependencyKey),
-                            projectPath);
-                    }
-
-                    var dependencyValue = dependencies.Value(dependencyKey);
-                    var dependencyTypeValue = LibraryDependencyType.Default;
-                    JsonString dependencyVersionAsString = null;
-
-                    if (dependencyValue is JsonObject)
-                    {
-                        // "dependencies" : { "Name" : { "version": "1.0", "type": "build" } }
-                        var dependencyValueAsObject = (JsonObject)dependencyValue;
-                        dependencyVersionAsString = dependencyValueAsObject.ValueAsString("version");
-
-                        // Remove support for flags (we only support build and nothing right now)
-                        var type = dependencyValueAsObject.ValueAsString("type");
-                        if (type != null)
-                        {
-                            dependencyTypeValue = LibraryDependencyType.Parse(type.Value);
-                        }
-                    }
-                    else if (dependencyValue is JsonString)
-                    {
-                        // "dependencies" : { "Name" : "1.0" }
-                        dependencyVersionAsString = (JsonString)dependencyValue;
-                    }
-                    else
-                    {
-                        throw FileFormatException.Create(
-                            string.Format("Invalid dependency version: {0}. The format is not recognizable.", dependencyKey),
-                            dependencyValue,
-                            projectPath);
-                    }
-
-                    SemanticVersionRange dependencyVersionRange = null;
-                    if (!string.IsNullOrEmpty(dependencyVersionAsString?.Value))
-                    {
-                        try
-                        {
-                            dependencyVersionRange = VersionUtility.ParseVersionRange(dependencyVersionAsString.Value);
-                        }
-                        catch (Exception ex)
-                        {
-                            throw FileFormatException.Create(
-                                ex,
-                                dependencyValue,
-                                projectPath);
-                        }
-                    }
-
-                    results.Add(new LibraryDependency
-                    {
-                        LibraryRange = new LibraryRange(dependencyKey, isGacOrFrameworkReference)
-                        {
-                            VersionRange = dependencyVersionRange,
-                            FileName = projectPath,
-                            Line = dependencyValue.Line,
-                            Column = dependencyValue.Column
-                        },
-                        Type = dependencyTypeValue
-                    });
-                }
-            }
-        }
-
-        private static bool TryGetStringEnumerable(JsonObject parent, string property, out IEnumerable<string> result)
-        {
-            var collection = new List<string>();
-            var valueInString = parent.ValueAsString(property);
-            if (valueInString != null)
-            {
-                collection.Add(valueInString);
-            }
-            else
-            {
-                var valueInArray = parent.ValueAsStringArray(property);
-                if (valueInArray != null)
-                {
-                    collection.AddRange(valueInArray);
-                }
-                else
-                {
-                    result = null;
-                    return false;
-                }
-            }
-
-            result = collection.SelectMany(value => value.Split(new[] { ' ', ',' }, StringSplitOptions.RemoveEmptyEntries));
-            return true;
-        }
-
-        public CompilerOptions GetCompilerOptions()
-        {
-            return _defaultCompilerOptions;
-        }
-
-        public CompilerOptions GetCompilerOptions(string configurationName)
-        {
-            CompilerOptions options;
-            if (_configurations.TryGetValue(configurationName, out options))
-            {
-                return options;
-            }
-
-            return null;
-        }
-
-        public CompilerOptions GetCompilerOptions(FrameworkName frameworkName)
-        {
-            CompilerOptions options;
-            if (_compilationOptions.TryGetValue(frameworkName, out options))
-            {
-                return options;
-            }
-
-            return null;
-        }
-
         public ICompilerOptions GetCompilerOptions(FrameworkName targetFramework,
-                                                  string configurationName)
+                                                   string configurationName)
         {
             // Get all project options and combine them
             var rootOptions = GetCompilerOptions();
@@ -489,188 +180,31 @@ namespace Microsoft.Dnx.Runtime
             return targetFrameworkInfo ?? _defaultTargetFrameworkConfiguration;
         }
 
-        private void BuildTargetFrameworksAndConfigurations(JsonObject projectJsonObject,
-            ICollection<DiagnosticMessage> diagnostics)
+        public CompilerOptions GetCompilerOptions()
         {
-            // Get the shared compilationOptions
-            _defaultCompilerOptions = GetCompilationOptions(projectJsonObject) ?? new CompilerOptions();
-
-            _defaultTargetFrameworkConfiguration = new TargetFrameworkInformation
-            {
-                Dependencies = new List<LibraryDependency>()
-            };
-
-            // Add default configurations
-            _configurations["Debug"] = new CompilerOptions
-            {
-                Defines = new[] { "DEBUG", "TRACE" },
-                Optimize = false
-            };
-
-            _configurations["Release"] = new CompilerOptions
-            {
-                Defines = new[] { "RELEASE", "TRACE" },
-                Optimize = true
-            };
-
-            // The configuration node has things like debug/release compiler settings
-            /*
-                {
-                    "configurations": {
-                        "Debug": {
-                        },
-                        "Release": {
-                        }
-                    }
-                }
-            */
-
-            var configurations = projectJsonObject.ValueAsJsonObject("configurations");
-            if (configurations != null)
-            {
-                foreach (var configKey in configurations.Keys)
-                {
-                    var compilerOptions = GetCompilationOptions(configurations.ValueAsJsonObject(configKey));
-
-                    // Only use this as a configuration if it's not a target framework
-                    _configurations[configKey] = compilerOptions;
-                }
-            }
-
-            // The frameworks node is where target frameworks go
-            /*
-                {
-                    "frameworks": {
-                        "net45": {
-                        },
-                        "k10": {
-                        }
-                    }
-                }
-            */
-
-            var frameworks = projectJsonObject.ValueAsJsonObject("frameworks");
-            if (frameworks != null)
-            {
-                foreach (var frameworkKey in frameworks.Keys)
-                {
-                    try
-                    {
-                        var frameworkToken = frameworks.ValueAsJsonObject(frameworkKey);
-                        var success = BuildTargetFrameworkNode(frameworkKey, frameworkToken);
-                        if (!success)
-                        {
-                            diagnostics?.Add(
-                                new DiagnosticMessage(
-                                    $"\"{frameworkKey}\" is an unsupported framework",
-                                    ProjectFilePath,
-                                    DiagnosticMessageSeverity.Error,
-                                    frameworkToken.Line,
-                                    frameworkToken.Column));
-                        }
-                    }
-                    catch (Exception ex)
-                    {
-                        throw FileFormatException.Create(ex, frameworks.Value(frameworkKey), ProjectFilePath);
-                    }
-                }
-            }
+            return _defaultCompilerOptions;
         }
 
-        /// <summary>
-        /// Parse a Json object which represents project configuration for a specified framework
-        /// </summary>
-        /// <param name="frameworkKey">The name of the framework</param>
-        /// <param name="frameworkValue">The Json object represent the settings</param>
-        /// <returns>Returns true if it successes.</returns>
-        private bool BuildTargetFrameworkNode(string frameworkKey, JsonObject frameworkValue)
+        public CompilerOptions GetCompilerOptions(string configurationName)
         {
-            // If no compilation options are provided then figure them out from the node
-            var compilerOptions = GetCompilationOptions(frameworkValue) ??
-                                  new CompilerOptions();
-
-            var frameworkName = FrameworkNameHelper.ParseFrameworkName(frameworkKey);
-
-            // If it's not unsupported then keep it
-            if (frameworkName == VersionUtility.UnsupportedFrameworkName)
+            CompilerOptions options;
+            if (_compilerOptionsByConfiguration.TryGetValue(configurationName, out options))
             {
-                // REVIEW: Should we skip unsupported target frameworks
-                return false;
+                return options;
             }
 
-            // Add the target framework specific define
-            var defines = new HashSet<string>(compilerOptions.Defines ?? Enumerable.Empty<string>());
-            var frameworkDefinition = Tuple.Create(frameworkKey, frameworkName);
-            var frameworkDefine = FrameworkNameHelper.MakeDefaultTargetFrameworkDefine(frameworkDefinition);
-
-            if (!string.IsNullOrEmpty(frameworkDefine))
-            {
-                defines.Add(frameworkDefine);
-            }
-
-            compilerOptions.Defines = defines;
-
-            var targetFrameworkInformation = new TargetFrameworkInformation
-            {
-                FrameworkName = frameworkName,
-                Dependencies = new List<LibraryDependency>()
-            };
-
-            var frameworkDependencies = new List<LibraryDependency>();
-
-            PopulateDependencies(
-                ProjectFilePath,
-                frameworkDependencies,
-                frameworkValue,
-                "dependencies",
-                isGacOrFrameworkReference: false);
-
-            var frameworkAssemblies = new List<LibraryDependency>();
-            PopulateDependencies(
-                ProjectFilePath,
-                frameworkAssemblies,
-                frameworkValue,
-                "frameworkAssemblies",
-                isGacOrFrameworkReference: true);
-
-            frameworkDependencies.AddRange(frameworkAssemblies);
-            targetFrameworkInformation.Dependencies = frameworkDependencies;
-
-            targetFrameworkInformation.WrappedProject = frameworkValue.ValueAsString("wrappedProject");
-
-            var binNode = frameworkValue.ValueAsJsonObject("bin");
-            if (binNode != null)
-            {
-                targetFrameworkInformation.AssemblyPath = binNode.ValueAsString("assembly");
-                targetFrameworkInformation.PdbPath = binNode.ValueAsString("pdb");
-            }
-
-            _compilationOptions[frameworkName] = compilerOptions;
-            _targetFrameworks[frameworkName] = targetFrameworkInformation;
-
-            return true;
+            return null;
         }
 
-        private static CompilerOptions GetCompilationOptions(JsonObject rawObject)
+        public CompilerOptions GetCompilerOptions(FrameworkName frameworkName)
         {
-            var rawOptions = rawObject.ValueAsJsonObject("compilationOptions");
-            if (rawOptions == null)
+            CompilerOptions options;
+            if (_compilerOptionsByFramework.TryGetValue(frameworkName, out options))
             {
-                return null;
+                return options;
             }
 
-            return new CompilerOptions
-            {
-                Defines = rawOptions.ValueAsStringArray("define"),
-                LanguageVersion = rawOptions.ValueAsString("languageVersion"),
-                AllowUnsafe = rawOptions.ValueAsNullableBoolean("allowUnsafe"),
-                Platform = rawOptions.ValueAsString("platform"),
-                WarningsAsErrors = rawOptions.ValueAsNullableBoolean("warningsAsErrors"),
-                Optimize = rawOptions.ValueAsNullableBoolean("optimize"),
-                KeyFile = rawOptions.ValueAsString("keyFile"),
-                DelaySign = rawOptions.ValueAsNullableBoolean("delaySign"),
-                StrongName = rawOptions.ValueAsNullableBoolean("strongName")
-            };
+            return null;
         }
     }
 }
